+++ monolith-tues-ex/src/main/java/com/whoiszxl/tues/trade/entity/vo/OmsInnerTransferVO.java
@@ -29,7 +29,7 @@ public class OmsInnerTransferVO extends AbstractObject implements Serializable {
     private Long id;
 
     @ApiModelProperty(value = "币种ID")
-    private Long coinId;
+    private Integer coinId;
 
     @ApiModelProperty(value = "转账金额")
     private BigDecimal amount;
